% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_categorical_columns.R
\name{explore_categorical_columns}
\alias{explore_categorical_columns}
\title{Performs EDA of categorical features.}
\usage{
explore_categorical_columns(df, categorical_col)
}
\arguments{
\item{df}{input data as a data frame}

\item{categorical_col}{vector containing categorical columns}
}
\value{
A data frame with details about unique, null values and most frequent category in every column and a vector containing plots
}
\description{
Creates a data frame containing column names and corresponding details about unique values, null values and most frequent category in every column
Plots count-plots for given categorical columns
Create a new factor from two existing factors, where the new factor's levels
are the union of the levels of the input factors.
}
\examples{
explore_categorical_columns(X, c('col1', 'col2'))
}
