% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_DBSCAN_clustering.R
\name{explore_DBSCAN_clustering}
\alias{explore_DBSCAN_clustering}
\title{fit and plot DBSCAN clustering on numeric columns of the dataset}
\usage{
explore_DBSCAN_clustering(df, eps = c(1), minPts = c(5))
}
\arguments{
\item{df}{dataframe: dataset}

\item{eps}{vector of the epsilon neighborhood size, optional, default to c(1)}

\item{minPts}{vector of number of minimum points in the eps region (for core points), default to c(5).}
}
\value{
a list of plots
}
\description{
fit and plot DBSCAN clustering on numeric columns of the dataset
}
\examples{
library(palmerpenguins)
explore_DBSCAN_clustering(penguins)
}
