% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_numeric_columns.R
\name{get_numeric_columns}
\alias{get_numeric_columns}
\title{get all numeric columns}
\usage{
get_numeric_columns(df)
}
\arguments{
\item{df}{data frame: the original dataset}
}
\value{
data frame: with all numeric columns from the original dataset
}
\description{
get all numeric columns
}
\examples{
library(palmerpenguins)
cols <- get_numeric_columns(penguins)
}
