% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_categorical_columns.R
\name{explore_categorical_columns}
\alias{explore_categorical_columns}
\title{Performs EDA of categorical features.}
\usage{
explore_categorical_columns(df, categorical_cols)
}
\arguments{
\item{df}{input data as a data frame}

\item{categorical_cols}{vector containing categorical columns}
}
\value{
A list object with first list element being a tibble with details about unique, null values and most frequent category in every column
and a second list element being count plots of user provided column names
}
\description{
Creates a data frame containing column names and corresponding details about unique values, null values and most frequent category in every column
Plots count-plots for given categorical columns
}
\examples{
library(dplyr)
library(MASS)
library(knitr)
df <- data.frame(lapply(survey[, c('Sex','Clap')], as.character), stringsAsFactors=FALSE) \%>\% tibble()
results <- explore_categorical_columns(df, c('Sex','Clap'))
results[[1]] \%>\% knitr::kable()
results[[2]][[1]]
results[[2]][[2]]
}
