% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_clustering.R
\name{explore_clustering}
\alias{explore_clustering}
\title{fit and plot K-Means, DBScan clustering algorithm on the dataset}
\usage{
explore_clustering(df, numeric_cols, hyperparameter_dict)
}
\arguments{
\item{df}{dataframe}

\item{numeric_cols}{vector of numeric columns' names}

\item{hyperparameter_dict}{dictionary of hyperparameters to be used}
}
\value{
a dictionary with each key = a clustering model name, value = list of plots generated by that model
}
\description{
fit and plot K-Means, DBScan clustering algorithm on the dataset
}
\examples{

}
