% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_DBSCAN_clustering.R
\name{explore_DBSCAN_clustering}
\alias{explore_DBSCAN_clustering}
\title{fit and plot DBSCAN clustering on numeric columns of the dataset}
\usage{
explore_DBSCAN_clustering(
  df,
  num_cols = NULL,
  eps = c(1),
  minPts = c(5),
  weights = NULL
)
}
\arguments{
\item{df}{dataframe: dataset}

\item{num_cols}{vector of numeric column names}

\item{eps}{vector of the epsilon neighborhood size}

\item{minPts}{vector of number of minimum points in the eps region (for core points). Default is c(5).}

\item{weights}{number: weights for the data points. Only needed to perform weighted clustering.}
}
\value{

}
\description{
fit and plot DBSCAN clustering on numeric columns of the dataset
}
