% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_KMeans_clustering.R
\name{explore_KMeans_clustering}
\alias{explore_KMeans_clustering}
\title{fits and plots KMeans clusterings on numeric columns of the dataset}
\usage{
explore_KMeans_clustering(
  df,
  centers = seq(from = 2, to = 10),
  iter.max = 10,
  algorithm = c("Hartigan-Wong", "Lloyd", "Forgy", "MacQueen")
)
}
\arguments{
\item{df}{dataframe}

\item{centers}{vector of number of clusters}

\item{iter.max}{integer: max number of iteration}

\item{algorithm}{character: may be abbreviated. Note that "Lloyd" and "Forgy" are alternative names for one algorithm}
}
\value{
a list of plots
}
\description{
fits and plots KMeans clusterings on numeric columns of the dataset
}
\examples{
\dontrun{
library(palmerpenguins)
explore_KMeans_clustering(penguins)
}
}
